Moved symlinks to new file. Removed dependency on pywin32. Using windows api for hard links
reorganization
fixed coverage
