+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 addopts = -p no:doctest --xdoctest --xdoctest-style=google
-norecursedirs = .git ignore build __pycache__ _rejected
+norecursedirs = .git ignore build __pycache__
 
 filterwarnings= default
                 ignore:.*No cfgstr given in Cacher constructor or call.*:Warning
