+++ requirements-win32.txt
@@ -1,3 +1,1 @@
-# pywin32 == 220
-# pypiwin32 == 219
 jaraco.windows
