+++ ubelt/util_platform.py
@@ -4,14 +4,11 @@ from os.path import exists
 from os.path import expanduser
 from os.path import join
 from os.path import normpath
-from os.path import islink
 import os
 import sys
 import pipes
 import six
-import warnings
 from ubelt import util_path
-from ubelt import util_io
 
 PY2 = sys.version_info.major == 2
 PY3 = sys.version_info.major == 3
@@ -20,9 +17,6 @@ WIN32  = sys.platform.startswith('win32')
 LINUX  = sys.platform.startswith('linux')
 DARWIN = sys.platform.startswith('darwin')
 POSIX = 'posix' in sys.builtin_module_names
-
-if WIN32:
-    from ubelt import _win32_links
 
 
 def platform_resource_dir():
@@ -217,178 +211,6 @@ def editfile(fpath, verbose=True):  # nocover
     if not exists(fpath):
         raise IOError('Cannot start nonexistant file: %r' % fpath)
     ub.cmd([editor, fpath], fpath, detatch=True)
-
-
-def symlink(real_path, link_path, overwrite=False, verbose=0):
-    """
-    Create a symbolic link.
-
-    Args:
-        path (str): path to real file or directory
-        link_path (str): path to desired location for symlink
-        overwrite (bool): overwrite existing symlinks.
-            This will not overwrite real files on systems with proper symlinks.
-            However, on older versions of windows junctions are
-            indistinguishable from real files, so we cannot make this
-            guarantee.  (default = False)
-        verbose (int):  verbosity level (default=0)
-
-    Notes:
-        There seems to be a corner case on Python2 and some versions of Windows
-        (whatever appveyor is using). However, my windows box works so idk.
-
-    Returns:
-        str: link path
-
-    CommandLine:
-        python -m ubelt.util_platform symlink:0
-
-    Example:
-        >>> import ubelt as ub
-        >>> dpath = ub.ensure_app_cache_dir('ubelt', 'test_symlink0')
-        >>> real_path = join(dpath, 'real_file.txt')
-        >>> link_path = join(dpath, 'link_file.txt')
-        >>> [ub.delete(p) for p in [real_path, link_path]]
-        >>> ub.writeto(real_path, 'foo')
-        >>> result = symlink(real_path, link_path)
-        >>> assert ub.readfrom(result) == 'foo'
-        >>> [ub.delete(p) for p in [real_path, link_path]]
-
-    Example:
-        >>> import ubelt as ub
-        >>> from os.path import dirname
-        >>> test_links = ub.import_module_from_path(dirname(__file__) + '/tests/test_links.py')
-        >>> dpath = ub.ensure_app_cache_dir('ubelt', 'test_symlink1')
-        >>> ub.delete(dpath)
-        >>> ub.ensuredir(dpath)
-        >>> test_links.dirstats(dpath)
-        >>> real_dpath = ub.ensuredir((dpath, 'real_dpath'))
-        >>> link_dpath = ub.augpath(real_dpath, base='link_dpath')
-        >>> real_path = join(dpath, 'afile.txt')
-        >>> link_path = join(dpath, 'afile.txt')
-        >>> [ub.delete(p) for p in [real_path, link_path]]
-        >>> ub.writeto(real_path, 'foo')
-        >>> result = symlink(real_dpath, link_dpath)
-        >>> assert ub.readfrom(link_path) == 'foo', 'read should be same'
-        >>> ub.writeto(link_path, 'bar')
-        >>> test_links.dirstats(dpath)
-        >>> assert ub.readfrom(link_path) == 'bar', 'very bad bar'
-        >>> assert ub.readfrom(real_path) == 'bar', 'changing link did not change real'
-        >>> ub.writeto(real_path, 'baz')
-        >>> test_links.dirstats(dpath)
-        >>> assert ub.readfrom(real_path) == 'baz', 'very bad baz'
-        >>> assert ub.readfrom(link_path) == 'baz', 'changing real did not change link'
-        >>> ub.delete(link_dpath, verbose=1)
-        >>> test_links.dirstats(dpath)
-        >>> assert not exists(link_dpath), 'link should not exist'
-        >>> assert exists(real_path), 'real path should exist'
-        >>> test_links.dirstats(dpath)
-        >>> ub.delete(dpath, verbose=1)
-        >>> test_links.dirstats(dpath)
-        >>> assert not exists(real_path)
-
-    TODO:
-        Can this be fixed on windows?
-        The main issue is that you need admin rights on Windows to symlink.
-    """
-    path = normpath(real_path)
-    link = normpath(link_path)
-    if verbose:
-        print('Creating symlink: path={} link={}'.format(path, link))
-    if islink(link):
-        if verbose:
-            print('symlink already exists')
-        if _readlink(link) == path:
-            if verbose > 1:
-                print('... and points to the right place')
-            return link
-        if verbose > 1:
-            if not exists(link):
-                print('... but it is broken and points somewhere else')
-            else:
-                print('... but it points somewhere else')
-        if overwrite:
-            util_io.delete(link, verbose > 1)
-
-    if WIN32:  # nocover
-        if exists(link) and not islink(link):
-            # On windows a broken link might still exist as a hard link or a
-            # junction. Overwrite it if it is a file and we cannot symlink.
-            # However, if it is a non-junction directory then do not overwrite
-            if verbose:
-                print('link location already exists')
-            is_junc = _win32_links._win32_is_junction(link)
-            # NOTE:
-            # in python2 broken junctions are directories and exist
-            # in python3 broken junctions are directories and do not exist
-            if os.path.isdir(link):
-                if is_junc:
-                    pointed = _win32_links._win32_read_junction(link)
-                    if path == pointed:
-                        if verbose:
-                            print('...and is a junction that points to the same place')
-                        return link
-                    else:
-                        if verbose:
-                            if not exists(pointed):
-                                print('...and is a broken junction that points somewhere else')
-                            else:
-                                print('...and is a junction that points somewhere else')
-                else:
-                    if verbose:
-                        print('...and is an existing real directory!')
-                    raise IOError('Cannot overwrite a real directory')
-
-            elif os.path.isfile(link):
-                if _win32_links._win32_is_hardlinked(link, path):
-                    if verbose:
-                        print('...and is a hard link that points to the same place')
-                    return link
-                else:
-                    if verbose:
-                        print('...and is a hard link that points somewhere else')
-                    if _can_symlink():
-                        raise IOError('Cannot overwrite potentially real file if we can symlink')
-            if overwrite:
-                if verbose:
-                    print('...overwriting')
-                util_io.delete(link, verbose > 1)
-            else:
-                if exists(link):
-                    raise IOError('Link already exists')
-
-        _win32_links._win32_symlink2(path, link, verbose=verbose)
-    else:
-        os.symlink(path, link)
-    return link
-
-
-def _readlink(link):
-    try:
-        return os.readlink(link)
-    except Exception:  # nocover
-        # On modern operating systems, we should never get here. (I think)
-        warnings.warn('Reading symlinks seems to not be supported')
-        raise
-
-
-def _can_symlink(verbose=0):  # nocover
-    """
-    Return true if we have permission to create real symlinks.
-    This check always returns True on non-win32 systems.  If this check returns
-    false, then we still may be able to use junctions.
-
-    CommandLine:
-        python -m ubelt.util_platform _can_symlink
-
-    Example:
-        >>> # Script
-        >>> print(_can_symlink(verbose=1))
-    """
-    if WIN32:
-        return _win32_links._win32_can_symlink(verbose)
-    else:
-        return True
 
 
 if __name__ == '__main__':
